Use indexed BLAST DB if already exists
Currently, when given the --proteins argument, prokka will automatically create a BLAST indexed database using the makeblastdb command, regardless of whether or not one already exists. This modification will search the directory of the protein database that is supplied by the user to determine whether an indexed database with the same prefix as the fasta file supplied already exists. If it does, prokka will use that instead of creating a new one.