Parents written to GFF before children
It is commonplace (but neither required, nor specification) for parent
features (e.g. `gene`) to precede their child features (e.g. `CDS`) in
GFF files.

This branch tweaks the sorting of features before writing to GFF, such
that parent features are written before their children. Note that this
is only applicable when `--addgenes` or `--addmrna` are invoked, and
this change is purely cosmetic.